test(snapshot): split iface/impl tests
Moves test classes to align with the production code package layout.
This enforces that we do not use package private access to
implementation classes in the interface tests.
