+++ snapshot/src/test/java/io/zeebe/snapshots/PersistedSnapshotTest.java
@@ -5,11 +5,11 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.zeebe.snapshots.impl;
+package io.zeebe.snapshots;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.zeebe.snapshots.ConstructableSnapshotStore;
+import io.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
 import io.zeebe.util.FileUtil;
 import io.zeebe.util.sched.testing.ActorSchedulerRule;
 import java.io.File;
