+++ snapshot/src/test/java/io/zeebe/snapshots/ReceivedSnapshotTest.java
@@ -5,18 +5,15 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.zeebe.snapshots.impl;
+package io.zeebe.snapshots;
 
 import static java.nio.file.StandardOpenOption.CREATE_NEW;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.zeebe.snapshots.ConstructableSnapshotStore;
-import io.zeebe.snapshots.PersistedSnapshot;
-import io.zeebe.snapshots.ReceivableSnapshotStore;
-import io.zeebe.snapshots.ReceivedSnapshot;
-import io.zeebe.snapshots.SnapshotChunk;
+import io.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
+import io.zeebe.snapshots.impl.InvalidSnapshotChecksum;
 import io.zeebe.util.FileUtil;
 import io.zeebe.util.sched.testing.ActorSchedulerRule;
 import java.io.IOException;
