+++ snapshot/src/test/java/io/zeebe/snapshots/SnapshotChunkWrapper.java
@@ -5,9 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.zeebe.snapshots.impl;
-
-import io.zeebe.snapshots.SnapshotChunk;
+package io.zeebe.snapshots;
 
 public final class SnapshotChunkWrapper implements SnapshotChunk {
