+++ snapshot/src/test/java/io/zeebe/snapshots/impl/FileBasedSnapshotChunkReaderTest.java
@@ -12,6 +12,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.zeebe.util.FileUtil;
 import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
@@ -60,11 +62,36 @@ public final class FileBasedSnapshotChunkReaderTest {
     assertThatThrownBy(reader::next).hasCauseInstanceOf(NoSuchFileException.class);
   }
 
+  @Test
+  public void shouldReadSnapshotChunks() throws IOException {
+    // given
+    try (final var snapshotChunkReader = newReader()) {
+      for (int i = 0; i < 2; i++) {
+        assertThat(snapshotChunkReader.hasNext()).isTrue();
+        // when
+        final var nextId = snapshotChunkReader.nextId();
+        final var chunk = snapshotChunkReader.next();
+
+        // then
+
+        assertThat(ByteBuffer.wrap(chunk.getChunkName().getBytes(StandardCharsets.UTF_8)))
+            .isNotNull()
+            .isEqualTo(nextId);
+        assertThat(chunk.getSnapshotId()).isEqualTo(snapshotDirectory.getFileName().toString());
+        assertThat(chunk.getTotalCount()).isEqualTo(2);
+        assertThat(chunk.getSnapshotChecksum()).isEqualTo(SNAPSHOT_CHECKSUM);
+        assertThat(chunk.getChecksum())
+            .isEqualTo(SnapshotChunkUtil.createChecksum(chunk.getContent()));
+        assertThat(snapshotDirectory.resolve(chunk.getChunkName()))
+            .hasBinaryContent(chunk.getContent());
+      }
+    }
+  }
+
   private FileBasedSnapshotChunkReader newReader() throws IOException {
     snapshotDirectory = temporaryFolder.getRoot().toPath();
     for (final var chunk : Arrays.asList("foo", "bar")) {
       final var path = snapshotDirectory.resolve(chunk);
-
       Files.createFile(path);
       Files.writeString(path, "content");
     }
